Act 4 cut scene
Act 4 cut scene. This should be implemented in a separate passage with a
display or goto from start, if crone reaches an endgame value (which
should be set somehow when there are not more tier 3 recipes to
discover). We need to capture a variable called "$last_unicorn", which
is the last tier 3 recipe discovered by the user, so that it can be used
in this passage.